pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Probirochniy <afl_damir@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grader = {path = "./grader"}
more-itertools = "^9.1.0"
requests = "^2.31.0"
types-requests = "^2.31.0.1"
dacite = "^1.8.1"
python-dotenv = "^1.0.0"
logger = {path = "./logger"}
vedis = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 60
